# src/main/resources/static/docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/tenants/onboard:
    post:
      tags:
      - tenant-controller
      operationId: onboardTenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantOnboardingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantOnboardingResponse"
...
  /api/v1/tenants/payment-settings:
    patch:
      tags:
      - tenant-controller
      operationId: updatePaymentSettings
      parameters:
      - name: X-Tenant-ID
        in: header
        required: true
        schema:
          type: string
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantPaymentSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /api/v1/tenants/{tenantId}:
    get:
      tags:
      - tenant-controller
      operationId: getTenant
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
      - name: X-Tenant-ID
        in: header
        required: true
        schema:
          type: string
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantOnboardingResponse"
...
components:
  schemas:
    TenantOnboardingRequest:
      required:
      - contactEmail
      - organizationName
      type: object
      properties:
        organizationName:
          maxLength: 255
          minLength: 0
          type: string
        contactEmail:
          type: string
    TenantOnboardingResponse:
      type: object
      properties:
        tenantId:
          type: string
        apiKey:
          type: string
        createdAt:
          type: string
          format: date-time
...
    TenantPaymentSettingsRequest:
      required:
      - yookassaSecretKey
      - yookassaShopId
      type: object
      properties:
        yookassaShopId:
          pattern: \d+
          type: string
        yookassaSecretKey:
          pattern: "^(live|test)_[a-zA-Z0-9\\-_]{32,}$"
          type: string
...